♻️ Refactor: #8 Three.js관련 파일 모듈화
