+++ play.html
@@ -13,7 +13,7 @@
   </head>
   <body>
     <main class="play-wrap">
-      <section>
+      <section class="play-cont">
         <div class="bottom-layout">
           <ul>
             <li>
@@ -38,6 +38,7 @@
             </li>
           </ul>
         </div>
+        <canvas class="play-canvas"></canvas>
       </section>
       <nav class="play-nav">
         <div>
@@ -119,5 +120,6 @@
       </aside>
     </main>
     <script src="/src/components/play.js" defer></script>
+    <script type="module" src="/src/components/origami.js" defer></script>
   </body>
 </html>
