+++ src/three/Camera.js
@@ -1,0 +1,11 @@
+import * as THREE from 'three';
+
+const camera = new THREE.PerspectiveCamera(
+  75,
+  window.innerWidth / window.innerHeight,
+  0.1,
+  1000
+);
+camera.position.z = 5;
+
+export { camera };
