+++ src/three/Lights.js
@@ -1,0 +1,7 @@
+import * as THREE from 'three';
+
+const ambientLight = new THREE.AmbientLight(0xffffff, 0.5);
+const directionalLight = new THREE.DirectionalLight(0xffffff, 0.5);
+directionalLight.position.set(1, 1, 1).normalize();
+
+export { ambientLight, directionalLight };
