+++ src/three/Plane.js
@@ -1,0 +1,53 @@
+import * as THREE from 'three';
+
+const colors = [
+  0x1aa24d, 0xf2482e, 0x343394, 0xfed53d, 0x000000, 0xff9fd5, 0x9ddef4,
+  0xff8158,
+];
+
+const getRandomColors = length => {
+  const frontColorIndex = Math.floor(Math.random() * length);
+  let backColorIndex = 0;
+
+  do {
+    backColorIndex = Math.floor(Math.random() * length);
+  } while (backColorIndex === frontColorIndex);
+
+  return [frontColorIndex, backColorIndex];
+};
+
+const [frontColorIndex, backColorIndex] = getRandomColors(colors.length);
+const frontColor = colors[frontColorIndex];
+const backColor = colors[backColorIndex];
+
+const geometry = new THREE.PlaneGeometry(2, 2);
+const material = new THREE.ShaderMaterial({
+  uniforms: {
+    colorFront: { value: new THREE.Color(frontColor) },
+    colorBack: { value: new THREE.Color(backColor) },
+  },
+  vertexShader: `
+  varying vec3 vNormal;
+  void main() {
+    vNormal = normalize(normalMatrix * normal);
+    gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
+  }
+`,
+  fragmentShader: `
+  uniform vec3 colorFront;
+  uniform vec3 colorBack;
+  varying vec3 vNormal;
+  void main() {
+    if (gl_FrontFacing) {
+      gl_FragColor = vec4(colorFront, 1.0);
+    } else {
+      gl_FragColor = vec4(colorBack, 1.0);
+    }
+  }
+`,
+  side: THREE.DoubleSide,
+});
+
+const plane = new THREE.Mesh(geometry, material);
+
+export { plane };
